changed way network learns about nodes
